### 2020-09-07 #2635 @marcosvital
Thank you for the suggestions, @Manolomon, I'll contact them.

In the meantime, you can work on a couple of suggestions, while we wait to hear from the potential reviewers:
* The paper should have a "statement of need" on a separate paragraph, with it's own header (this came to be a norm on all JOSS papers recently).
* The repository need to show "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" (this will be one point on [reviewers' checklist](https://joss.readthedocs.io/en/latest/review_checklist.html), and I didn't see information about this as I took a look at the repository).
