### 2020-09-07 ca64e35
Create CODE_OF_CONDUCT.md
