+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+# Contributing
+
+When contributing to this repository, please first discuss the change you wish to make via issue,
+email, or any other method with the owners of this repository before making a change.
+
+Please note we have a [code of conduct](https://github.com/Manolomon/uvehavior-desktop/blob/master/CODE_OF_CONDUCT.md), please follow it in all your interactions with the project.
+
+## Pull Request Process
+
+Please refer to the [Pull Request Template](https://github.com/Manolomon/uvehavior-desktop/blob/master/CODE_OF_CONDUCT.md) in order to make changes to this repository.
