### 2020-09-22 #2692 @marcosvital
@SteveViss and @62442katieb: thank you again for for accepting review this submission for JOSS.

Even if you are not starting the review right now, please accept the invite, as it has an expiration date (there is a link under Reviewer instructions & questions and you should also get an email notification). Furthermore, please check the instructions and checklists above, and let me know if you need any assistance.

You can also tag @Manolomon if you need to ask specific questions about the submission or to address any changes that might be necessary in the submitted paper or in the repository.
