### 2020-11-05 #2692 @62442katieb
In the mean time, I have a few small comments regarding the paper:
- In the *Summary*, the authors should consider switching mentioning ethology as "behavioral analysis" and "behavioral science" are broad, nonspecific terms, and specifying their scope would be helpful in introducing the tool.
- The authors should provide a description how this task is currently performed, in addressing the "state of the field". From an outside perspective, do people usually use another software tool? Pen and paper? Excel spreadsheets? I think it might help, too, in showing the utility and value of their tool.

Regarding documentation:
- The authors should consider adding to the wiki homepage, as it is not immediately apparent to users how to find the information they're looking for, once they've been directed there from the README. Even just mentioning in the main body text the different pages (Home, Maintenance, Test, Usage) and a brief description would be really helpful to orient users.
- Additionally, the "Usage" page does not contain much usage on how to set up and perform an experiment, skipping from "tests and subject of interest must be added. Every element regarding experiment planning and management are displayed in the following view..." to "Once the behavioral observations have been recorded..." without much information about how one goes about adding subjects and texts or how to add recordings or bind keys to behaviors. I think users would be much more capable of making full use of this tool to improve their research if a few more details are added to the documentation.
- Installation instructions from building from source are good, but there's no information beyond links for installing from a disk image, .exe, or app image. A few instructions might be helpful here, too.
