+++ paper.md
@@ -42,8 +42,7 @@ bibliography: paper.bib
 
 # Summary
 
-Understanding behavior is a relevant task, detecting patterns in how living organisms interact and respond to different stimuli means an area of interest capable of influencing complementary activities in psychology, nutrition, and pharmacology. Research in behavioral science is, however, a high-dimensional problem that involves several activities, from the conceptualization and planning of the study to the analysis of the collected results. Currently, there are several automated software to behavioral analysis, however, they are expensive and limited to the analysis of particular behaviors that only the human eye can identify.
-In this regard, we introduce UVehavior, and open-source tool that allows researchers to manage, conduct, and compare behavioral observations in several animals and even the human being.
+Understanding behavior is a relevant task, detecting patterns in how living organisms interact and respond to different physical and emotional stimuli is an essential tool in several areas of the science, including ethology, psychology, nutrition, and pharmacology, among others. Particularly, studies in behavioral pharmacology involve a high-dimensional issue that involves several activities, from the conceptualization and planning of the study, the identification and selection of specific behaviors to the analysis of the collected data. Currently, there are several automated software to analyze behavioral patterns that permit identify emotional, affective, and motor changes associated with physiological, psychological or pharmacological manipulations, however, they are expensive and limited to the analysis of particular behaviors that only the human eye can identify. In this regard, we introduce UVehavior, and open-source tool that allows researchers to manage, conduct, and compare behavioral observations in several animals and even the human being; which represent a free tool to researcher interested in experimental analysis of behavior.
 
 # Statement of Need
 
@@ -60,6 +59,8 @@ In ethology, the behavior was typically categorized by trained observers taking 
 While conducting the experiment, every subject is recorded during each defined test, where the occurrence of a specific behavior must be logged.
 
 ![Annotation main view. The left side of the window shows the playing video and controls, while the right side displays the evaluation's data and the behaviors' log.\label{fig:annotation}](src/assets/img/annotation.png)
+
+The analysis of behavior using video annotations is a complicated task. In this way, the simultaneous recording of diverse behaviors with makeshift tools, such as video editors, may be imprecise and time-consuming. Additionally, they introduce human error in the recording of duration of each behavior, so the annotation of parallel observations is not possible, and they may be restrictive in the management and export of data to specific software to the analysis. Currently, there are different software to automatized record of the animal behavior, however, most of them are very expensive and complicated to use, because they were developed for the analysis of behavior of different species under particular experimental conditions. In spite of that software are capable of evaluating behavioral patterns automatically, unfortunately they are not accurate and need strict supervision by researcher. So, it is necessary to have a software tool specially designed to record the behavior in an easy and practical way.
 
 UVehavior assists in this core activity by featuring a suite where the researcher can evaluate their subjects, tests, and behaviors to observe, in one of these test recordings.
 The desktop-tool, plays the test video, while the observer analyzes the behavior of the subject. Therefore, every time the subject performs a specific behavior, the researcher logs it by pressing one key previously bound to a behavior.
