+++ paper.md
@@ -42,14 +42,14 @@ bibliography: paper.bib
 
 # Summary
 
-Understanding behavior is a relevant task, detecting patterns in how living organisms interact and respond to different physical and emotional stimuli is an essential tool in several areas of the science, including ethology, psychology, nutrition, and pharmacology, among others. Particularly, studies in behavioral pharmacology involve a high-dimensional issue that involves several activities, from the conceptualization and planning of the study, the identification and selection of specific behaviors to the analysis of the collected data. Currently, there are several automated software to analyze behavioral patterns that permit identify emotional, affective, and motor changes associated with physiological, psychological or pharmacological manipulations, however, they are expensive and limited to the analysis of particular behaviors that only the human eye can identify. In this regard, we introduce UVehavior, and open-source tool that allows researchers to manage, conduct, and compare behavioral observations in several animals and even the human being; which represent a free tool to researcher interested in experimental analysis of behavior.
+Understanding behavior is a relevant task, detecting patterns in how living organisms interact and respond to different physical and emotional stimuli is an essential tool in several areas of the science, including ethology, psychology, nutrition, and pharmacology, among others. Particularly, the study of behavior in pharmacology is a high-dimensional topic that involves several activities, from the conceptualization and planning of the study, the identification and selection of specific behaviors to the analysis of the collected data. Currently, there are several automated software tools to analyze behavioral patterns that permit identify emotional, affective and motor changes associated with physiological, psychological, or pharmacological manipulations; however, they are expensive and limited to the analysis of particular behaviors that only the human eye can identify. In this regard, we introduce UVehavior, and open-source tool that allows researchers to manage, conduct, and compare behavioral observations in several animals and even human beings; which represent a free tool for researchers interested in the experimental analysis of behavior.
 
 # Statement of Need
 
 According to @Dugatkin:2020 behavior is the coordinated responses of whole living organisms to internal and/or external stimuli, considering behavior as an expression of the activity of the nervous system.
 Neuroethology is an important area of neurosciences involved in the study of the neurobiological, neurochemical, and neurophysiological bases underlying behavior [@Bassler:1984]. Therefore, this science is interested in the quantification of the behavioral changes through the measure of frequency, latency and time duration of specific behavior under normal or altered states induced by drugs, experimental manipulations, or the experience of the individual.
 
-On the other hand, before beginning a study on behavior, it is necessary to consider whether the design of the experiment is adequate to allow causal and logical explanations, orienting the investigation towards the adequate data interpretation. The experimental design considers the study subject, conditions of the environment and the control of the variables studied [@Kirk:2012].
+However, before beginning a study on behavior, it is necessary to consider whether the design of the experiment is adequate to allow causal and logical explanations, orienting the investigation towards adequate data interpretation. The experimental design considers the study subject, conditions of the environment and the control of the variables studied [@Kirk:2012].
 
 As part of the planning stage, the researcher defines and distributes the groups in which the subjects of study are going to be evaluated. Is by their belonging group that the individuals are subjected to a treatment, for further analysis of the effects on their behavior [@Kirk:2012]. UVehavior allows the researcher to keep track of their subjects, and distribute them by groups, as well as managing the tests to be conducted. In this regard, every test involves a specific amount of time, and a variety of behaviors to be analyzed.
 
@@ -60,7 +60,7 @@ While conducting the experiment, every subject is recorded during each defined t
 
 ![Annotation main view. The left side of the window shows the playing video and controls, while the right side displays the evaluation's data and the behaviors' log.\label{fig:annotation}](src/assets/img/annotation.png)
 
[…]
