### 2021-02-22 #2692 @62442katieb
Additionally, in reviewing the paper, I noticed a few typos and grammatical issues:
- page 1, line 39: "**Is** by their belonging group that the individuals are subjected to a treatment...", should read, "**It is** by their..."
- page 2, line 41: "UVehavior allows the researcher to **keep** track of their subjects, and **distribute** them by groups, as well as **managing** the tests to be conducted." The verb tenses in this sentence are mixed, to retain parallel tenses, it should read, "UVehavior allows the researcher to **keep** track of their subjects, **distribute** them by groups, and **manage** the tests to be conducted." (minor additional edit of clunky conjunctions)
