### 2021-02-22 #2692 @Manolomon
Hello @62442katieb, thanks for your review. I'm addressing the typos as soon as possible. Regarding the documentation, we are currently working on it as part of an upcoming update to the system. We hope to include most of those changes this week.

Thank you, and best regards
