+++ src/app/core/services/electron/electron.service.ts
@@ -2,12 +2,13 @@ import { Injectable } from '@angular/core';
 
 // If you import a module but never use any of the imported values other than as TypeScript types,
 // the resulting javascript file will look as if you never imported the module at all.
-import { ipcRenderer, webFrame, remote } from 'electron';
+import { ipcRenderer, webFrame } from 'electron';
+import * as remote from '@electron/remote';
 import * as childProcess from 'child_process';
 import * as fs from 'fs';
 
 @Injectable({
-  providedIn: 'root',
+  providedIn: 'root'
 })
 export class ElectronService {
   ipcRenderer: typeof ipcRenderer;
@@ -25,13 +26,16 @@ export class ElectronService {
     if (this.isElectron) {
       this.ipcRenderer = window.require('electron').ipcRenderer;
       this.webFrame = window.require('electron').webFrame;
-      this.remote = window.require('electron').remote;
+
+      // If you want to use remote object in renderer process, please set enableRemoteModule to true in main.ts
+      // this.remote = window.require('@electron/remote');
+      // console.log('remote - globalShortcut', this.remote.globalShortcut);
 
       this.childProcess = window.require('child_process');
       this.fs = window.require('fs');
     }
   }
-
+  
   toggleMediaPlayback(status: boolean) {
     remote.getCurrentWindow().webContents.audioMuted = status;
   }
