### 2021-06-12 #2692 @Manolomon
Hello again @marcosvital. I'm sorry this revision has taken so long. During the last months, I had been dealing with all my graduation-related processes and couldn't find the time to keep maintaining uvehavior. Now I and my colleagues have updated the repository and most importantly, generated full-length documentation on the [wiki site](https://github.com/Manolomon/uvehavior-desktop/wiki), as requested by @62442katieb.

I'm sorry again for all the late responses, but I'm now fully committed to the project and the revision.
