### 2021-06-18 #2692 @62442katieb
Hi @marcosvital and @Manolomon ! I'm happy to continue reviewing, but have a conference and my defense in the next two weeks, so I probably won't get to it until the first week of July. Looking forward to sifting through the updated documentation!
