### 2021-09-08 #2692 @marcosvital
Thank you, @heacosta. I'll generate a new version of the manuscript.

Dear @SteveViss and @62442katieb, thank you for your patience and all the work you've done so far. I'm generating an updated proof and you can also check the updated documentation referred by @Manolomon above. Can you proceed with the reviewing process from there?
