### 2021-09-23 #2692 @62442katieb
I just finished looking over the revised manuscript, documentation, and software. My comments are below:

Manuscript:
- line 12: "... we introduce UVehavior, **and** open-source tool..." I believe the bold "and" should be "an"
- line 75: "... whereas the horizontal one represents the _chronological order_..."; it would be more accurate to say that the horizontal axis represents time, as it presents elapsed seconds (which is important and valuable), instead of an ordinal scale of behavior occurrence

Documentation:
-  Overall very helpful! Thanks for taking the time to write all this out, it made using the tool much easier.
- Throughout the manuscript and in the GitHub README, the tool's name is written as "UVehavior", but in the documentation it is often "Uvehavior". The authors might want to edit this for consistency.
- There are a few typos in the documentation, the authors should consider proofreading and editing this text.

Software:
- Looks and works great, overall! Very nice tool.
- Authors might want to consider allowing users to resize the video viewer.
