### 2022-03-11 #2692 @Manolomon
Hello everyone.

Over the past year, my active participation in the development of the project entitled [Uvhevior](https://github.com/Manolomon/uvehavior-desktop) has been on hold and is not comparable to what it was at the time this submission started.
Unfortunately, this difference has escalated to a point that hinders the review from proceeding.
Due to this, my co-authors have come to the decision to **withdraw this submission** from being considered at the Journal of Open Source Software.

I sincerely apologize for all the inconvenience that this issue had caused, especially to the reviewers and the editorial team.
